ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.6.4
